+++ modules/v1_14_R1/src/com/dsh105/echopet/compat/nms/v1_14_R1/entity/ai/PetGoalLookAtPlayer.java
@@ -19,45 +19,44 @@ package com.dsh105.echopet.compat.nms.v1_14_R1.entity.ai;
 import com.dsh105.echopet.compat.api.ai.APetGoalLookAtPlayer;
 import com.dsh105.echopet.compat.api.ai.PetGoalType;
 import com.dsh105.echopet.compat.nms.v1_14_R1.entity.EntityPet;
-
 import net.minecraft.server.v1_14_R1.Entity;
 import net.minecraft.server.v1_14_R1.EntityHuman;
 import net.minecraft.server.v1_14_R1.IEntitySelector;
 
 @SuppressWarnings("rawtypes")
 public class PetGoalLookAtPlayer extends APetGoalLookAtPlayer{
-
+	
 	private EntityPet pet;
 	protected Entity player;
 	private float range;
 	private int ticksLeft;
 	private float chance;
 	private Class clazz;
-
+	
 	public PetGoalLookAtPlayer(EntityPet pet, Class c){
 		this.pet = pet;
 		this.range = 8.0F;
 		this.chance = 0.02F;
 		this.clazz = c;
 	}
-
+	
 	public PetGoalLookAtPlayer(EntityPet pet, Class c, float f, float f1){
 		this.pet = pet;
 		this.range = f;
 		this.chance = f1;
 		this.clazz = c;
 	}
-
+	
 	@Override
 	public PetGoalType getType(){
 		return PetGoalType.TWO;
 	}
-
+	
 	@Override
 	public String getDefaultKey(){
 		return "LookAtPlayer";
 	}
-
+	
 	@SuppressWarnings("unchecked")
 	@Override
 	public boolean shouldStart(){
@@ -67,29 +66,29 @@ public class PetGoalLookAtPlayer extends APetGoalLookAtPlayer{
 			return false;
 		}else{
 			if(this.clazz == EntityHuman.class){
-				this.player = this.pet.world.findNearbyPlayer(this.pet, (double) this.range);
+				this.player = this.pet.world.findNearbyPlayer(this.pet, this.range);
 			}else{
-				this.player = (Entity) this.pet.world.a(this.clazz, this.pet.getBoundingBox().grow((double) this.range, 3.0D, (double) this.range), IEntitySelector.b(pet)).stream().findAny().orElse(null);
+				this.player = (Entity) this.pet.world.a(this.clazz, this.pet.getBoundingBox().grow(this.range, 3.0D, this.range), IEntitySelector.b(pet)).stream().findAny().orElse(null);
 			}
 			return this.player != null;
 		}
 	}
-
+	
 	@Override
 	public boolean shouldContinue(){
 		return this.player.isAlive() && (this.pet.h(this.player) <= (double) (this.range * this.range) && this.ticksLeft > 0);
 	}
-
+	
 	@Override
 	public void start(){
 		this.ticksLeft = 40 + this.pet.random().nextInt(40);
 	}
-
+	
 	@Override
 	public void finish(){
 		this.player = null;
 	}
-
+	
 	@Override
 	public void tick(){
 		// 1.9: this.a.getControllerLook().a(this.b.locX, this.b.locY + this.b.getHeadHeight(), this.b.locZ, this.a.cE(), this.a.N());
